refactor: move signature methods to v1
